requirements.txt
chainlit>=1.1.0
PyMuPDF>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
